Move settings conversion into TidepoolManager and switch to direct uploads
Move all therapy settings conversion logic from the standalone
TrioSettingsAdapter into a BaseTidepoolManager extension, eliminating
the separate injectable class.

Replace observer-based upload pattern (CarbRatios, InsulinSensitivities,
BasalProfile, BGTargets, PumpSettings observers) with direct
Task.detached upload calls in each state model's save(), matching the
existing Nightscout upload pattern. Keep SettingsObserver and
PreferencesObserver for closedLoop/units/algorithm preference changes.
